Move 'key interface' to its own file, update initial values
key_interface will act as the abstraction layer between the GUI and the
sn76489. It's purpose is to emulate a 'key' register, so that notes can
more easily be interacted with.

The key register is emulated by setting the attenuation to 15 whenever
the channel is not keyed.

The default registers for the four channels are also updated:
 - The tone channels are initialized with values for a C major chord
 - The noise channel is initialized to white noise