SG-1000 support
This patch adds support for running on the SG-1000/SC-3000 using Mode-0.

In `SMSlib_compat.h`, a number of defines are made so that the code can
continue to mostly use SMSlib function calls.

Mode-0 is used, rather than the more capable Mode-2, in the hope that if
this can be made to load over tape it will be a bit faster to load with
the small colour table.

Code has also changed to use `_loadTileMap` instead of `_loadTileMapArea`,
as `SG_loadTileMapArea` appeared to have a VRAM-corrupting bug.

Keyboard patterns are split into separate files for the outline, active
keys, and inactive keys. This is to generate separate defines for each
block of patterns. There may be an unknown number of all-zero padding
patterns between each colour-change.