Customized SGlib for tape support
Normally, SGlib relies on interrupts to trigger input reading and waiting
for vblank. However, when running from tape, the interrupt vector belongs
to the BASIC cartridge ROM.

Instead, we need to poll the VDP status register to determine when we've
reached vblank. To do this, the following changes are made:

 * Removed code that re-enables interrupts after VDP accesses
 * Adjusted SG_waitForVBlank to poll the status register
 * Moved key input reading to SG_waitForVBlank

With these changes, so long as the program uses SG_waitForVBlank every
frame, this should work as a drop-in replacement.