Build SN76489 TestRom for SC-3000 tape
This change adds tape builds for the TestRom, to run from BASIC IIIa or
BASIC IIIb cartridge.

The changes to get this working are:
 - Linker parameters to stet the new code and data locations
 - Link against the customized crt0_sg
 - Link against the customized SGlib

For now, ihx2sms is used, but this pads the file to a near round number,
making it bigger than needed. The unused space below 0x9800 is removed
using dd, but the excess size at the end is currently left in place. A
different tool may be needed if we cross the size boundary and generate
another 16k page.

SC-TapeWave is used to convert the binary into a wave file that can be
played into the SC-3000's cassette input.

To run the program:

  1. `LOAD`
  2. play the audio file
  3. `CALL &H9800`